Change package name to io.github.dnalchemist.spi.mapstruct to avoid conflicts with the original project
