+++ pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>no.entur.mapstruct.spi</groupId>
+    <groupId>io.github.dnalchemist.mapstruct.spi</groupId>
     <artifactId>spi-protobuf-parent</artifactId>
     <version>1.47-SNAPSHOT</version>
     <packaging>pom</packaging>
