+++ spi-impl/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>no.entur.mapstruct.spi</groupId>
+        <groupId>io.github.dnalchemist.mapstruct.spi</groupId>
         <artifactId>spi-protobuf-parent</artifactId>
         <version>1.47-SNAPSHOT</version>
     </parent>
@@ -40,7 +40,7 @@
                         <configuration>
                             <compilerArgument>-proc:none</compilerArgument>
                             <includes>
-                                <include>no/entur/mapstruct/spi/protobuf/ProcessingEnvOptionsHolder.java</include>
+                                <include>io/github/dnalchemist/mapstruct/spi/protobuf/ProcessingEnvOptionsHolder.java</include>
                             </includes>
                         </configuration>
                     </execution>
