+++ spi-impl/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-no.entur.mapstruct.spi.protobuf.ProcessingEnvOptionsHolder
+io.github.dnalchemist.mapstruct.spi.protobuf.ProcessingEnvOptionsHolder
