+++ usage/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>no.entur.mapstruct.spi</groupId>
+        <groupId>io.github.dnalchemist.mapstruct.spi</groupId>
         <artifactId>spi-protobuf-parent</artifactId>
         <version>1.47-SNAPSHOT</version>
     </parent>
@@ -46,7 +46,7 @@
                     <annotationProcessorPaths>
                         <path>
                             <!-- Add your SPI implementation. -->
-                            <groupId>no.entur.mapstruct.spi</groupId>
+                            <groupId>io.github.dnalchemist.mapstruct.spi</groupId>
                             <artifactId>protobuf-spi-impl</artifactId>
                             <version>${project.version}</version>
                             <!-- In real life, this should be a fixed version from a different
@@ -59,7 +59,7 @@
                         </path>
                     </annotationProcessorPaths>
                     <compilerArgs>
-                        <arg>-AmapstructSpi.enumPostfixOverrides=no.entur.mapstruct.example.EnumPostfixOverrideProtos=INVALID</arg>
+                        <arg>-AmapstructSpi.enumPostfixOverrides=io.github.dnalchemist.mapstruct.example.EnumPostfixOverrideProtos=INVALID</arg>
                     </compilerArgs>
                 </configuration>
             </plugin>
