+++ usage/src/main/java/io/github/dnalchemist/mapstruct/spi/protobuf/House.java
@@ -1,4 +1,4 @@
-package no.entur.mapstruct.spi.protobuf;
+package io.github.dnalchemist.mapstruct.spi.protobuf;
 
 /*-
  * #%L
@@ -23,15 +23,15 @@ package no.entur.mapstruct.spi.protobuf;
  * #L%
  */
 
-public class MultiNumber {
-	Number number;
+public class House {
+	private String name;
 
-	public Number getNumber() {
-		return number;
+	public String getName() {
+		return name;
 	}
 
-	public void setNumber(Number number) {
-		this.number = number;
+	public void setName(String name) {
+		this.name = name;
 	}
 
 }
