+++ usage/src/main/java/io/github/dnalchemist/mapstruct/spi/protobuf/Status.java
@@ -1,4 +1,4 @@
-package no.entur.mapstruct.spi.protobuf;
+package io.github.dnalchemist.mapstruct.spi.protobuf;
 
 /*-
  * #%L
@@ -23,15 +23,7 @@ package no.entur.mapstruct.spi.protobuf;
  * #L%
  */
 
-public class House {
-	private String name;
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
+public enum Status {
+	STARTED,
+	STOPPED
 }
