+++ usage/src/main/proto/User.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "no.entur.mapstruct.example";
+option java_package = "io.github.dnalchemist.mapstruct.example";
 option java_outer_classname = "UserProtos";
 
 enum EnumStatus {
